Refactor disc.py into seperate files
