--- dissect/disc/c_udf.py
@@ -1,7 +1,0 @@
-UDF_MAGICS = [
-    # These are actually magics from ECMA-167, but there do not seem to be many other filesystems other than UDF that
-    # use ECMA-167
-    b"BEA01",
-    b"BOOT2",
-    b"TEA01",
-]
