+++ dissect/disc/iso/c_iso_9660.py
