+++ dissect/disc/iso/c_rockridge.py
