+++ tests/test_hybrid.py
@@ -4,9 +4,11 @@ from typing import BinaryIO
 
 import pytest
 
-from dissect.disc.disc import DISC, ISOFormat, log
+from dissect.disc.base import DiscFormat
+from dissect.disc.disc import DISC, log
 from dissect.disc.exceptions import FileNotFoundError
-
+from dissect.disc.iso.iso9660 import ISO9660Disc
+from dissect.disc.iso.rockridge import RockridgeDisc
 
 LONG_FILENAME = "100_character_long_filename_" + ("a" * 68) + ".txt"
 
@@ -16,31 +18,38 @@ log.setLevel(logging.DEBUG)
 
 @pytest.mark.parametrize("use_path_table", [False, True])
 def test_hybrid(hybrid_iso: BinaryIO, use_path_table: bool, caplog) -> None:
-    rockridge_fs = DISC(hybrid_iso)
+    rockridge_disc = DISC(hybrid_iso)
+    rockridge_fs = rockridge_disc.fs
     # Assert defaulting to rockridge
-    assert rockridge_fs.iso_format == ISOFormat.ROCKRIDGE
+    assert rockridge_disc.selected_format == DiscFormat.ROCKRIDGE
+    assert isinstance(rockridge_fs, RockridgeDisc)
 
-    joliet_fs = DISC(hybrid_iso, preference=ISOFormat.JOLIET)
+    joliet_disc = DISC(hybrid_iso, preference=DiscFormat.JOLIET)
+    joliet_fs = joliet_disc.fs
 
     # Assert warning raised when selecting Joliet
+    assert joliet_disc.selected_format == DiscFormat.JOLIET
+    assert isinstance(joliet_fs, ISO9660Disc)
     assert "Treating disc as Joliet even though Rockridge is available" in caplog.text
-    assert joliet_fs.iso_format == ISOFormat.JOLIET
 
-    plain_fs = DISC(hybrid_iso, preference=ISOFormat.PLAIN)
-    assert plain_fs.iso_format == ISOFormat.PLAIN
+    iso9660_disc = DISC(hybrid_iso, preference=DiscFormat.ISO9660)
+    iso9660_fs = iso9660_disc.fs
+
+    assert iso9660_disc.selected_format == DiscFormat.ISO9660
+    assert isinstance(iso9660_fs, ISO9660Disc)
 
     contents = b"My full filename should be supported on Joliet"
 
     assert joliet_fs.get(LONG_FILENAME, use_path_table).open().read() == contents
     assert joliet_fs.get(LONG_FILENAME, use_path_table).open().read() == contents
-    assert plain_fs.get("100_CHAR.TXT", use_path_table).open().read() == contents
+    assert iso9660_fs.get("100_CHAR.TXT", use_path_table).open().read() == contents
 
 
 @pytest.mark.parametrize(
     "fs_format,expected",
     [
         (
-            ISOFormat.JOLIET,
+            DiscFormat.JOLIET,
             dict(
                 {
                     "/": 33,
@@ -51,7 +60,7 @@ def test_hybrid(hybrid_iso: BinaryIO, use_path_table: bool, caplog) -> None:
             ),
         ),
         (
-            ISOFormat.ROCKRIDGE,
+            DiscFormat.ROCKRIDGE,
             dict(
                 {
                     "/": 28,
@@ -62,7 +71,7 @@ def test_hybrid(hybrid_iso: BinaryIO, use_path_table: bool, caplog) -> None:
             ),
         ),
         (
-            ISOFormat.PLAIN,
+            DiscFormat.ISO9660,
             dict(
                 {
                     "/": 28,
@@ -74,39 +83,41 @@ def test_hybrid(hybrid_iso: BinaryIO, use_path_table: bool, caplog) -> None:
         ),
     ],
 )
-def test_path_table(hybrid_iso: BinaryIO, fs_format: ISOFormat, expected: dict) -> None:
-    fs = DISC(hybrid_iso, preference=fs_format)
-    assert fs.path_table == expected
+def test_path_table(hybrid_iso: BinaryIO, fs_format: DiscFormat, expected: dict) -> None:
+    disc = DISC(hybrid_iso, preference=fs_format)
+
+    assert isinstance(disc.fs, ISO9660Disc)
+    assert disc.fs.path_table == expected
 
 
-@pytest.mark.parametrize("fs_format", [ISOFormat.JOLIET, ISOFormat.ROCKRIDGE, ISOFormat.PLAIN])
-def test_primary_volume_descriptor(hybrid_iso: BinaryIO, fs_format: ISOFormat):
-    fs = DISC(hybrid_iso, preference=fs_format)
-    assert fs.volume_name == "CDROM"
-    assert fs.primary_volume.application_id.decode(fs.name_encoding).startswith("GENISOIMAGE ISO 9660")
-    assert fs.primary_volume.system_id.decode(fs.name_encoding).startswith("LINUX")
+@pytest.mark.parametrize("fs_format", [DiscFormat.JOLIET, DiscFormat.ROCKRIDGE, DiscFormat.ISO9660])
+def test_primary_volume_descriptor(hybrid_iso: BinaryIO, fs_format: DiscFormat):
+    disc = DISC(hybrid_iso, preference=fs_format)
+    assert disc.fs.name == "CDROM"
+    assert disc.fs.primary_volume.application_id.decode(disc.fs.encoding).startswith("GENISOIMAGE ISO 9660")
+    assert disc.fs.primary_volume.system_id.decode(disc.fs.encoding).startswith("LINUX")
 
 
-@pytest.mark.parametrize("fs_format", [ISOFormat.JOLIET, ISOFormat.ROCKRIDGE, ISOFormat.PLAIN])
+@pytest.mark.parametrize("fs_format", [DiscFormat.JOLIET, DiscFormat.ROCKRIDGE, DiscFormat.ISO9660])
 @pytest.mark.parametrize("use_path_table", [False, True])
-def test_notfound(hybrid_iso: BinaryIO, fs_format: ISOFormat, use_path_table: bool):
-    fs = DISC(hybrid_iso, preference=fs_format)
+def test_notfound(hybrid_iso: BinaryIO, fs_format: DiscFormat, use_path_table: bool):
+    disc = DISC(hybrid_iso, preference=fs_format)
     with pytest.raises(FileNotFoundError):
-        fs.get("a/does_not_exists.txt", use_path_table)
+        disc.fs.get("a/does_not_exists.txt", use_path_table)
 
 
 @pytest.mark.parametrize(
     "fs_format,filename",
     [
-        (ISOFormat.JOLIET, LONG_FILENAME),
-        (ISOFormat.ROCKRIDGE, LONG_FILENAME),
-        (ISOFormat.PLAIN, "100_CHAR.TXT"),
+        (DiscFormat.JOLIET, LONG_FILENAME),
+        (DiscFormat.ROCKRIDGE, LONG_FILENAME),
+        (DiscFormat.ISO9660, "100_CHAR.TXT"),
     ],
 )
 @pytest.mark.parametrize("use_path_table", [False, True])
-def test_metadata(hybrid_iso: BinaryIO, fs_format: ISOFormat, filename: str, use_path_table: bool):
-    fs = DISC(hybrid_iso, preference=fs_format)
-    entry = fs.get(filename, use_path_table)
+def test_metadata(hybrid_iso: BinaryIO, fs_format: DiscFormat, filename: str, use_path_table: bool):
+    disc = DISC(hybrid_iso, preference=fs_format)
+    entry = disc.fs.get(filename, use_path_table)
 
     # All three properties are the same for this file, which makes it easy to test :)
     assert entry.mtime == datetime(2024, 3, 9, 12, 40, 4, tzinfo=timezone(timedelta(seconds=3600)))
